ALX HIGH LEVEL

+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -10,7 +10,7 @@ void hash_table_print(const hash_table_t *ht)
 {
 	hash_node_t *current = NULL;
 	int comma = 0;
-	unsigned long int i = 0;
+	int i = 0;
 
 	if (!ht)
 	{
